refactored Context.jsx into separate Context files to focus on each concern, making it easier to maintain and scalable
